move dependency management and service configs into bases
