+++ bases/renku_data_services/data_api/app.py
@@ -9,7 +9,6 @@ from sanic_ext.extras.validation.validators import VALIDATION_ERROR
 from ulid import ULID
 
 from renku_data_services import errors
-from renku_data_services.app_config import Wiring
 from renku_data_services.base_api.error_handler import CustomErrorHandler
 from renku_data_services.base_api.misc import MiscBP
 from renku_data_services.base_models.core import Slug
@@ -23,6 +22,7 @@ from renku_data_services.crc.blueprints import (
     ResourcePoolUsersBP,
     UserResourcePoolsBP,
 )
+from renku_data_services.data_api.dependencies import DependencyManager
 from renku_data_services.data_connectors.blueprints import DataConnectorsBP
 from renku_data_services.message_queue.blueprints import MessageQueueBP
 from renku_data_services.namespace.blueprints import GroupsBP
@@ -64,7 +64,7 @@ def _patched_validate_body(
         ) from e
 
 
-def register_all_handlers(app: Sanic, wiring: Wiring) -> Sanic:
+def register_all_handlers(app: Sanic, dm: DependencyManager) -> Sanic:
     """Register all handlers on the application."""
     # WARNING: The regex is not actually used in most cases, instead the conversion function must raise a ValueError
     app.router.register_pattern("ulid", ULID.from_str, r"^[0-7][0-9A-HJKMNP-TV-Z]{25}$")
@@ -74,192 +74,190 @@ def register_all_handlers(app: Sanic, wiring: Wiring) -> Sanic:
     resource_pools = ResourcePoolsBP(
         name="resource_pools",
         url_prefix=url_prefix,
-        rp_repo=wiring.rp_repo,
-        authenticator=wiring.authenticator,
-        user_repo=wiring.user_repo,
+        rp_repo=dm.rp_repo,
+        authenticator=dm.authenticator,
+        user_repo=dm.user_repo,
     )
-    classes = ClassesBP(name="classes", url_prefix=url_prefix, repo=wiring.rp_repo, authenticator=wiring.authenticator)
+    classes = ClassesBP(name="classes", url_prefix=url_prefix, repo=dm.rp_repo, authenticator=dm.authenticator)
     quota = QuotaBP(
         name="quota",
         url_prefix=url_prefix,
-        rp_repo=wiring.rp_repo,
-        authenticator=wiring.authenticator,
-        quota_repo=wiring.quota_repo,
+        rp_repo=dm.rp_repo,
+        authenticator=dm.authenticator,
+        quota_repo=dm.quota_repo,
     )
-    users = KCUsersBP(name="users", url_prefix=url_prefix, repo=wiring.kc_user_repo, authenticator=wiring.authenticator)
+    users = KCUsersBP(name="users", url_prefix=url_prefix, repo=dm.kc_user_repo, authenticator=dm.authenticator)
     user_secrets = UserSecretsBP(
         name="user_secrets",
         url_prefix=url_prefix,
-        secret_repo=wiring.user_secrets_repo,
-        authenticator=wiring.authenticator,
+        secret_repo=dm.user_secrets_repo,
+        authenticator=dm.authenticator,
     )
     resource_pools_users = ResourcePoolUsersBP(
         name="resource_pool_users",
         url_prefix=url_prefix,
-        repo=wiring.user_repo,
-        authenticator=wiring.authenticator,
-        kc_user_repo=wiring.kc_user_repo,
+        repo=dm.user_repo,
+        authenticator=dm.authenticator,
+        kc_user_repo=dm.kc_user_repo,
     )
     user_resource_pools = UserResourcePoolsBP(
         name="user_resource_pools",
         url_prefix=url_prefix,
-        repo=wiring.user_repo,
-        authenticator=wiring.authenticator,
-        kc_user_repo=wiring.kc_user_repo,
+        repo=dm.user_repo,
+        authenticator=dm.authenticator,
+        kc_user_repo=dm.kc_user_repo,
     )
-    clusters = ClustersBP(
-        name="clusters", url_prefix=url_prefix, repo=wiring.cluster_repo, authenticator=wiring.authenticator
-    )
+    clusters = ClustersBP(name="clusters", url_prefix=url_prefix, repo=dm.cluster_repo, authenticator=dm.authenticator)
     storage = StorageBP(
         name="storage",
         url_prefix=url_prefix,
-        storage_repo=wiring.storage_repo,
-        authenticator=wiring.gitlab_authenticator,
+        storage_repo=dm.storage_repo,
+        authenticator=dm.gitlab_authenticator,
     )
     storage_schema = StorageSchemaBP(name="storage_schema", url_prefix=url_prefix)
     user_preferences = UserPreferencesBP(
         name="user_preferences",
         url_prefix=url_prefix,
-        user_preferences_repo=wiring.user_preferences_repo,
-        authenticator=wiring.authenticator,
+        user_preferences_repo=dm.user_preferences_repo,
+        authenticator=dm.authenticator,
     )
-    misc = MiscBP(name="misc", url_prefix=url_prefix, apispec=wiring.spec, version=wiring.version)
+    misc = MiscBP(name="misc", url_prefix=url_prefix, apispec=dm.spec, version=dm.version)
     project = ProjectsBP(
         name="projects",
         url_prefix=url_prefix,
-        project_repo=wiring.project_repo,
-        project_member_repo=wiring.project_member_repo,
-        authenticator=wiring.authenticator,
-        user_repo=wiring.kc_user_repo,
-        session_repo=wiring.session_repo,
-        data_connector_repo=wiring.data_connector_repo,
-        project_migration_repo=wiring.project_migration_repo,
-        metrics=wiring.metrics,
+        project_repo=dm.project_repo,
+        project_member_repo=dm.project_member_repo,
+        authenticator=dm.authenticator,
+        user_repo=dm.kc_user_repo,
+        session_repo=dm.session_repo,
+        data_connector_repo=dm.data_connector_repo,
+        project_migration_repo=dm.project_migration_repo,
+        metrics=dm.metrics,
     )
     project_session_secrets = ProjectSessionSecretBP(
         name="project_session_secrets",
         url_prefix=url_prefix,
-        session_secret_repo=wiring.project_session_secret_repo,
-        authenticator=wiring.authenticator,
+        session_secret_repo=dm.project_session_secret_repo,
+        authenticator=dm.authenticator,
     )
     group = GroupsBP(
         name="groups",
         url_prefix=url_prefix,
-        authenticator=wiring.authenticator,
-        group_repo=wiring.group_repo,
-        metrics=wiring.metrics,
+        authenticator=dm.authenticator,
+        group_repo=dm.group_repo,
+        metrics=dm.metrics,
     )
     session_environments = EnvironmentsBP(
         name="session_environments",
         url_prefix=url_prefix,
-        session_repo=wiring.session_repo,
-        authenticator=wiring.authenticator,
+        session_repo=dm.session_repo,
+        authenticator=dm.authenticator,
     )
     session_launchers = SessionLaunchersBP(
         name="sessions_launchers",
         url_prefix=url_prefix,
-        session_repo=wiring.session_repo,
-        authenticator=wiring.authenticator,
-        metrics=wiring.metrics,
+        session_repo=dm.session_repo,
+        authenticator=dm.authenticator,
+        metrics=dm.metrics,
     )
     builds = (
         BuildsBP(
             name="builds",
             url_prefix=url_prefix,
-            session_repo=wiring.session_repo,
-            authenticator=wiring.authenticator,
+            session_repo=dm.session_repo,
+            authenticator=dm.authenticator,
         )
-        if wiring.config.builds.enabled
+        if dm.config.builds.enabled
         else None
     )
     oauth2_clients = OAuth2ClientsBP(
         name="oauth2_clients",
         url_prefix=url_prefix,
-        connected_services_repo=wiring.connected_services_repo,
-        authenticator=wiring.authenticator,
+        connected_services_repo=dm.connected_services_repo,
+        authenticator=dm.authenticator,
     )
     oauth2_connections = OAuth2ConnectionsBP(
         name="oauth2_connections",
         url_prefix=url_prefix,
-        connected_services_repo=wiring.connected_services_repo,
-        authenticator=wiring.authenticator,
-        internal_gitlab_authenticator=wiring.gitlab_authenticator,
+        connected_services_repo=dm.connected_services_repo,
+        authenticator=dm.authenticator,
+        internal_gitlab_authenticator=dm.gitlab_authenticator,
     )
     repositories = RepositoriesBP(
         name="repositories",
         url_prefix=url_prefix,
-        git_repositories_repo=wiring.git_repositories_repo,
-        authenticator=wiring.authenticator,
-        internal_gitlab_authenticator=wiring.gitlab_authenticator,
+        git_repositories_repo=dm.git_repositories_repo,
+        authenticator=dm.authenticator,
+        internal_gitlab_authenticator=dm.gitlab_authenticator,
     )
     notebooks = NotebooksBP(
         name="notebooks_old",
         url_prefix=url_prefix,
-        authenticator=wiring.authenticator,
-        nb_config=wiring.nb_config,
-        internal_gitlab_authenticator=wiring.gitlab_authenticator,
-        git_repo=wiring.git_repositories_repo,
-        rp_repo=wiring.rp_repo,
-        user_repo=wiring.kc_user_repo,
-        storage_repo=wiring.storage_repo,
+        authenticator=dm.authenticator,
+        nb_config=dm.nb_config,
+        internal_gitlab_authenticator=dm.gitlab_authenticator,
+        git_repo=dm.git_repositories_repo,
+        rp_repo=dm.rp_repo,
+        user_repo=dm.kc_user_repo,
+        storage_repo=dm.storage_repo,
     )
     notebooks_new = NotebooksNewBP(
         name="notebooks",
         url_prefix=url_prefix,
-        authenticator=wiring.authenticator,
-        nb_config=wiring.nb_config,
-        project_repo=wiring.project_repo,
-        project_session_secret_repo=wiring.project_session_secret_repo,
-        session_repo=wiring.session_repo,
-        storage_repo=wiring.storage_repo,
-        rp_repo=wiring.rp_repo,
-        user_repo=wiring.kc_user_repo,
-        data_connector_repo=wiring.data_connector_repo,
-        data_connector_secret_repo=wiring.data_connector_secret_repo,
-        internal_gitlab_authenticator=wiring.gitlab_authenticator,
-        metrics=wiring.metrics,
+        authenticator=dm.authenticator,
+        nb_config=dm.nb_config,
+        project_repo=dm.project_repo,
+        project_session_secret_repo=dm.project_session_secret_repo,
+        session_repo=dm.session_repo,
+        storage_repo=dm.storage_repo,
+        rp_repo=dm.rp_repo,
+        user_repo=dm.kc_user_repo,
+        data_connector_repo=dm.data_connector_repo,
+        data_connector_secret_repo=dm.data_connector_secret_repo,
+        internal_gitlab_authenticator=dm.gitlab_authenticator,
+        metrics=dm.metrics,
     )
     platform_config = PlatformConfigBP(
         name="platform_config",
         url_prefix=url_prefix,
-        platform_repo=wiring.platform_repo,
-        authenticator=wiring.authenticator,
+        platform_repo=dm.platform_repo,
+        authenticator=dm.authenticator,
     )
     message_queue = MessageQueueBP(
         name="search",
         url_prefix=url_prefix,
-        authenticator=wiring.authenticator,
-        session_maker=wiring.config.db.async_session_maker,
-        reprovisioning_repo=wiring.reprovisioning_repo,
-        user_repo=wiring.kc_user_repo,
-        group_repo=wiring.group_repo,
-        project_repo=wiring.project_repo,
-        authz=wiring.authz,
+        authenticator=dm.authenticator,
+        session_maker=dm.config.db.async_session_maker,
+        reprovisioning_repo=dm.reprovisioning_repo,
+        user_repo=dm.kc_user_repo,
+        group_repo=dm.group_repo,
+        project_repo=dm.project_repo,
+        authz=dm.authz,
     )
     search = SearchBP(
         name="search2",
         url_prefix=url_prefix,
-        authenticator=wiring.authenticator,
+        authenticator=dm.authenticator,
         search_reprovision=SearchReprovision(
-            search_updates_repo=wiring.search_updates_repo,
-            reprovisioning_repo=wiring.reprovisioning_repo,
-            solr_config=wiring.config.solr,
-            user_repo=wiring.kc_user_repo,
-            group_repo=wiring.group_repo,
-            project_repo=wiring.project_repo,
-            data_connector_repo=wiring.data_connector_repo,
+            search_updates_repo=dm.search_updates_repo,
+            reprovisioning_repo=dm.reprovisioning_repo,
+            solr_config=dm.config.solr,
+            user_repo=dm.kc_user_repo,
+            group_repo=dm.group_repo,
+            project_repo=dm.project_repo,
+            data_connector_repo=dm.data_connector_repo,
         ),
-        solr_config=wiring.config.solr,
-        authz=wiring.authz,
-        metrics=wiring.metrics,
+        solr_config=dm.config.solr,
+        authz=dm.authz,
+        metrics=dm.metrics,
     )
     data_connectors = DataConnectorsBP(
         name="data_connectors",
         url_prefix=url_prefix,
-        data_connector_repo=wiring.data_connector_repo,
-        data_connector_secret_repo=wiring.data_connector_secret_repo,
-        authenticator=wiring.authenticator,
-        metrics=wiring.metrics,
+        data_connector_repo=dm.data_connector_repo,
+        data_connector_secret_repo=dm.data_connector_secret_repo,
+        authenticator=dm.authenticator,
+        metrics=dm.metrics,
     )
     app.blueprint(
         [
