+++ components/renku_data_services/app_config/__init__.py
@@ -1,3 +1,1 @@
 """Configuration parsing utilities."""
-
-from renku_data_services.app_config.config import Wiring  # noqa: F401
