+++ components/renku_data_services/secrets/config.py
@@ -4,24 +4,13 @@ import os
 import secrets
 from dataclasses import dataclass, field
 from pathlib import Path
-from typing import Any, Self
+from typing import Self
 
 from cryptography.hazmat.primitives import serialization
 from cryptography.hazmat.primitives.asymmetric import rsa
 from cryptography.hazmat.primitives.asymmetric.types import PrivateKeyTypes
-from jwt import PyJWKClient
-from yaml import safe_load
 
-import renku_data_services.secrets
-from renku_data_services import base_models, errors
-from renku_data_services.app_config.config import KeycloakConfig
-from renku_data_services.authn.dummy import DummyAuthenticator
-from renku_data_services.authn.keycloak import KeycloakAuthenticator
-from renku_data_services.db_config.config import DBConfig
-from renku_data_services.k8s.client_interfaces import K8sCoreClientInterface
-from renku_data_services.k8s.clients import DummyCoreClient, K8sCoreClient
-from renku_data_services.secrets.db import LowLevelUserSecretsRepo
-from renku_data_services.utils.core import oidc_discovery
+from renku_data_services import errors
 
 
 @dataclass
@@ -100,81 +89,3 @@ class PrivateSecretsConfig:
             raise errors.ConfigurationError(message="Old secret service private key is not an RSAPrivateKey")
 
         return cls(private_key=private_key, previous_private_key=previous_private_key)
-
-
-@dataclass
-class Config:
-    """Main config for secrets service."""
-
-    db: DBConfig
-    secrets: PrivateSecretsConfig
-    keycloak: KeycloakConfig | None
-    app_name: str = "secrets_storage"
-    version: str = "0.0.1"
-    dummy_stores: bool = False
-    spec: dict[str, Any] = field(init=False, default_factory=dict)
-
-    def __post_init__(self) -> None:
-        spec_file = Path(renku_data_services.secrets.__file__).resolve().parent / "api.spec.yaml"
-        with open(spec_file) as f:
-            self.spec = safe_load(f)
-
-    @classmethod
-    def from_env(cls) -> Self:
-        """Load values from environment."""
-        dummy_stores = os.environ.get("DUMMY_STORES", "false").lower() == "true"
-        db = DBConfig.from_env()
-        secrets_config = PrivateSecretsConfig.from_env()
-        version = os.environ.get("VERSION", "0.0.1")
-        keycloak = None
-        if not dummy_stores:
-            KeycloakConfig.from_env()
-
-        return cls(db=db, secrets=secrets_config, version=version, keycloak=keycloak)
-
-
-@dataclass
-class Wiring:
-    """Wiring for secrets service."""
-
-    authenticator: base_models.Authenticator
-    config: Config
-    core_client: K8sCoreClientInterface
-    _user_secrets_repo: LowLevelUserSecretsRepo | None = field(default=None, repr=False, init=False)
-
-    @property
-    def user_secrets_repo(self) -> LowLevelUserSecretsRepo:
-        """The DB adapter for users."""
-        if not self._user_secrets_repo:
-            self._user_secrets_repo = LowLevelUserSecretsRepo(
-                session_maker=self.config.db.async_session_maker,
-            )
-        return self._user_secrets_repo
-
-    @classmethod
-    def from_env(cls) -> "Wiring":
-        """Create a config from environment variables."""
-        authenticator: base_models.Authenticator
-        core_client: K8sCoreClientInterface
-        config = Config.from_env()
-
-        if config.dummy_stores:
-            authenticator = DummyAuthenticator()
-            core_client = DummyCoreClient({}, {})
-        else:
-            assert config.keycloak is not None
-            oidc_disc_data = oidc_discovery(config.keycloak.keycloak_url, config.keycloak.keycloak_realm)
-            jwks_url = oidc_disc_data.get("jwks_uri")
-            if jwks_url is None:
-                raise errors.ConfigurationError(
-                    message="The JWKS url for Keycloak cannot be found from the OIDC discovery endpoint."
-                )
-            jwks = PyJWKClient(jwks_url)
-            authenticator = KeycloakAuthenticator(jwks=jwks, algorithms=config.keycloak.algorithms)
-            core_client = K8sCoreClient()
-
-        return cls(
-            config=config,
-            authenticator=authenticator,
-            core_client=core_client,
-        )
