+++ test/bases/renku_data_services/data_api/test_repositories.py
@@ -8,7 +8,7 @@ import pytest_asyncio
 from sanic import Sanic
 from sanic_testing.testing import SanicASGITestClient
 
-from renku_data_services.app_config import Wiring
+from renku_data_services.app_config import DependencyManager
 from renku_data_services.connected_services.dummy_async_oauth2_client import DummyAsyncOAuth2Client
 from renku_data_services.data_api.app import register_all_handlers
 from renku_data_services.migrations.core import run_migrations_for_app
@@ -16,7 +16,7 @@ from test.utils import SanicReusableASGITestClient
 
 
 @pytest_asyncio.fixture(scope="session")
-async def oauth2_test_client_setup(app_config: Wiring) -> SanicASGITestClient:
+async def oauth2_test_client_setup(app_config: DependencyManager) -> SanicASGITestClient:
     app_config.async_oauth2_client_class = DummyAsyncOAuth2Client
     app_config.connected_services_repo.async_oauth2_client_class = DummyAsyncOAuth2Client
     app = Sanic(app_config.app_name)
