+++ test/components/renku_data_services/authz/test_authorization.py
@@ -8,7 +8,7 @@ from authzed.api.v1 import (
 )
 from ulid import ULID
 
-from renku_data_services.app_config import Wiring
+from renku_data_services.app_config import DependencyManager
 from renku_data_services.authz.authz import _AuthzConverter
 from renku_data_services.authz.models import Member, Role, Scope, Visibility
 from renku_data_services.base_models import APIUser
@@ -26,7 +26,7 @@ regular_user2 = APIUser(is_admin=False, id="user2-id", access_token="some-token2
 
 
 @pytest_asyncio.fixture
-async def bootstrap_admins(app_config_instance: Wiring, event_loop) -> None:
+async def bootstrap_admins(app_config_instance: DependencyManager, event_loop) -> None:
     run_migrations_for_app("common")
     authz = app_config_instance.authz
     admins = [admin_user]
@@ -45,7 +45,9 @@ async def bootstrap_admins(app_config_instance: Wiring, event_loop) -> None:
 
 @pytest.mark.asyncio
 @pytest.mark.parametrize("public_project", [True, False])
-async def test_adding_deleting_project(app_config_instance: Wiring, bootstrap_admins, public_project: bool) -> None:
+async def test_adding_deleting_project(
+    app_config_instance: DependencyManager, bootstrap_admins, public_project: bool
+) -> None:
     project_owner = regular_user1
     assert project_owner.id
     authz = app_config_instance.authz
@@ -95,7 +97,7 @@ async def test_adding_deleting_project(app_config_instance: Wiring, bootstrap_ad
 @pytest.mark.parametrize("public_project", [True, False])
 @pytest.mark.parametrize("granted_role", [Role.VIEWER, Role.EDITOR, Role.OWNER])
 async def test_granting_access(
-    app_config_instance: Wiring, bootstrap_admins, public_project: bool, granted_role: Role
+    app_config_instance: DependencyManager, bootstrap_admins, public_project: bool, granted_role: Role
 ) -> None:
     project_owner = regular_user1
     assert project_owner.id
@@ -140,7 +142,9 @@ async def test_granting_access(
 
 @pytest.mark.asyncio
 @pytest.mark.parametrize("public_project", [True, False])
-async def test_listing_users_with_access(app_config_instance: Wiring, public_project: bool, bootstrap_admins) -> None:
+async def test_listing_users_with_access(
+    app_config_instance: DependencyManager, public_project: bool, bootstrap_admins
+) -> None:
     project_owner = regular_user1
     assert project_owner.id
     assert regular_user2.id
@@ -186,7 +190,7 @@ async def test_listing_users_with_access(app_config_instance: Wiring, public_pro
 
 
 @pytest.mark.asyncio
-async def test_listing_projects_with_access(app_config_instance: Wiring, bootstrap_admins) -> None:
+async def test_listing_projects_with_access(app_config_instance: DependencyManager, bootstrap_admins) -> None:
     authz = app_config_instance.authz
     public_project_id = ULID()
     private_project_id1 = ULID()
@@ -307,7 +311,7 @@ async def test_listing_projects_with_access(app_config_instance: Wiring, bootstr
 
 
 @pytest.mark.asyncio
-async def test_listing_non_public_projects(app_config_instance: Wiring, bootstrap_admins) -> None:
+async def test_listing_non_public_projects(app_config_instance: DependencyManager, bootstrap_admins) -> None:
     authz = app_config_instance.authz
     public_project_id = ULID()
     private_project_id1 = ULID()
