+++ test/components/renku_data_services/db/test_sqlalchemy_user_preferences_repo.py
@@ -3,7 +3,7 @@ from hypothesis import HealthCheck, given, settings, target
 
 import renku_data_services.base_models as base_models
 from renku_data_services import errors
-from renku_data_services.app_config import Wiring
+from renku_data_services.app_config import DependencyManager
 from renku_data_services.migrations.core import run_migrations_for_app
 from test.components.renku_data_services.user_preferences_models.hypothesis import (
     project_slug_strat,
@@ -16,7 +16,7 @@ from test.utils import create_user_preferences
 @settings(suppress_health_check=[HealthCheck.function_scoped_fixture], deadline=None)
 @pytest.mark.asyncio
 async def test_user_preferences_insert_get(
-    project_slug: str, app_config_instance: Wiring, loggedin_user: base_models.APIUser
+    project_slug: str, app_config_instance: DependencyManager, loggedin_user: base_models.APIUser
 ) -> None:
     run_migrations_for_app("common")
     user_preferences_repo = app_config_instance.user_preferences_repo
@@ -30,7 +30,7 @@ async def test_user_preferences_insert_get(
 @settings(suppress_health_check=[HealthCheck.function_scoped_fixture], deadline=None, max_examples=25)
 @pytest.mark.asyncio
 async def test_user_preferences_add_pinned_project(
-    project_slugs: list[str], app_config_instance: Wiring, loggedin_user: base_models.APIUser
+    project_slugs: list[str], app_config_instance: DependencyManager, loggedin_user: base_models.APIUser
 ) -> None:
     run_migrations_for_app("common")
     target(len(project_slugs))
@@ -53,7 +53,7 @@ async def test_user_preferences_add_pinned_project(
 @settings(suppress_health_check=[HealthCheck.function_scoped_fixture], deadline=None, max_examples=25)
 @pytest.mark.asyncio
 async def test_user_preferences_add_pinned_project_existing(
-    project_slugs: list[str], app_config_instance: Wiring, loggedin_user: base_models.APIUser
+    project_slugs: list[str], app_config_instance: DependencyManager, loggedin_user: base_models.APIUser
 ) -> None:
     run_migrations_for_app("common")
     target(len(project_slugs))
@@ -77,7 +77,7 @@ async def test_user_preferences_add_pinned_project_existing(
 @settings(suppress_health_check=[HealthCheck.function_scoped_fixture], deadline=None, max_examples=25)
 @pytest.mark.asyncio
 async def test_user_preferences_delete_pinned_project(
-    project_slugs: list[str], app_config_instance: Wiring, loggedin_user: base_models.APIUser
+    project_slugs: list[str], app_config_instance: DependencyManager, loggedin_user: base_models.APIUser
 ) -> None:
     run_migrations_for_app("common")
     target(len(project_slugs))
@@ -102,7 +102,7 @@ async def test_user_preferences_delete_pinned_project(
 @settings(suppress_health_check=[HealthCheck.function_scoped_fixture], deadline=None, max_examples=25)
 @pytest.mark.asyncio
 async def test_user_preferences_add_pinned_project_respects_maximum(
-    project_slugs: list[str], app_config_instance: Wiring, loggedin_user: base_models.APIUser
+    project_slugs: list[str], app_config_instance: DependencyManager, loggedin_user: base_models.APIUser
 ) -> None:
     run_migrations_for_app("common")
     target(len(project_slugs))
